repaired tests
