+++ src/main/java/edu/iis/mto/search/MockForSequenceSearcher.java
@@ -2,12 +2,12 @@ package edu.iis.mto.search;
 
 public class MockForSequenceSearcher implements SequenceSearcher {
 
+    private int counter =0;
     @Override public SearchResult search(int elem, int[] seq) {
-        SearchResult.Builder object = SearchResult.builder();
-        for(int i=0;i<seq.length;i++){
-            if(seq[i]==elem)
-                return object.withFound(true).withPosition(i).build();
-        }
-        return object.withFound(false).withPosition(-1).build();
+        counter = counter +1;
+        return SearchResult.builder().build();
+    }
+    public int getCounter(){
+        return counter;
     }
 }
